Statically inject Timeline Profiler hooks
We shouldn't use meta programming like spread that gets deopted.
